+++ project-3/src/db/operators/physical/extended/SortOperator.java
@@ -21,5 +21,5 @@ public interface SortOperator extends Operator, UnaryNode<Operator>, SeekableOpe
      *
      * @return The index of the tuple returned by the last call to <pre>getNextTuple()</pre> or -1 if no such tuple exists.
      */
-    long getTupleIndex();
+    int getTupleIndex();
 }
